+++ MordoEngine/src/Renderer/AreaSelectorRenderer.h
@@ -19,10 +19,8 @@ private:
 
 public:
 	AreaSelectorRenderer(Shader& shader, float radius, int segments);
-	void Render(const glm::mat4* view,
-			    const glm::mat4* projection,
-				const glm::mat4* model) override;
-	void SetHeights(const std::vector<float>& heights, const glm::vec3& position) override;
+	void Render(const RenderContext& renderContext) override;
+	void SetHeights(const std::vector<float>& heights, const glm::vec3& position);
 	void SetRadio(float radius);
 	~AreaSelectorRenderer();
 };