+++ MordoEngine/src/Renderer/RenderContext.h
@@ -1,6 +1,7 @@
 #pragma once
 #include <glm/glm.hpp>
 
+
 struct RenderContext
 {
 	glm::mat4 projection{ 1.0f };